Crate spade

Spade

Delaunay triangulations for the rust eco system.

Features

A 2D Delaunay triangulation: DelaunayTriangulation

Uses exact geometric predicate evaluation, preventing construction errors due to precision loss.

A 2D constrained Delaunay triangulation: ConstrainedDelaunayTriangulation

Supports vertex removal

Serde support with the serde feature.

Modules

handles

Handle types used for traversal and modification of triangulations.

internals

Internals that must be published due to technical reasons. This is not the place you are looking for. A change to these items is not considered to be a breaking change.

iterators

Iterators over various elements of Delaunay triangulations.

Structs

CdtEdge

Undirected edge type of a ConstrainedDelaunayTriangulation (CDT).

ConstrainedDelaunayTriangulation

A two dimensional constrained Delaunay triangulation.

DelaunayTriangulation

A two dimensional Delaunay triangulation.

LastUsedVertexHintGenerator

A hint generator that returns the last used vertex as hint.

LineSideInfo

Describes on which side of a line a a point lies.

Point2

A two dimensional point.

PointProjection

Indicates a point’s projected position relative to an edge.

Enums

InsertionError

The error type used for inserting elements into a triangulation.

PositionInTriangulation

Describes a position in a triangulation.

Constants

MAX_ALLOWED_VALUE

The largest allowed coordinate value that can be inserted into Delaunay triangulations. This value is equal to 2201.

MIN_ALLOWED_VALUE

The smallest allowed coordinate value greater than zero that can be inserted into Delaunay triangulations. This value is equal to 2-142.

Traits

HasPosition

An object with position.

HintGenerator

A structure used to speed up common operations on delaunay triangulations like insertion and geometry queries by providing hints on where to start searching for elements.

SpadeNum

A coordinate type that can be used with a triangulation.

Triangulation

Defines common operations on triangulations.

Functions

mitigate_underflow

Prevents underflow issues of a position by setting any coordinate that is too small to zero.

validate_coordinate

Checks if a coordinate value is suitable for insertion into a Delaunay triangulation.

validate_vertex

Checks if a vertex is suitable for insertion into a Delaunay triangulation.

Type Definitions

HierarchyHintGenerator

A hint generator based on a hierarchy of triangulations optimized for randomly accessing elements of the triangulation.