Enum spade::delaunay::PositionInTriangulation

pub enum PositionInTriangulation<V: Copy, F: Copy, E: Copy> {
    InTriangle(F),
    OutsideConvexHull(E),
    OnPoint(V),
    OnEdge(E),
    NoTriangulationPresent,
}

Stores information about a point's position in triangulation.

Used as a return type of DelaunayTriangulation::locate(..).

Variants
InTriangle(F)

The point is contained in a triangle.

OutsideConvexHull(E)

The point is outside the convex hull. The given edge is an edge that is close to the queried position.

OnPoint(V)

A vertex with this position has already been inserted. Its handle is given.

OnEdge(E)

The point lies on an edge.

NoTriangulationPresent

There is no valid triangulation yet, thus, less than two points where inserted.