Struct spade::delaunay::DelaunayTriangulation

pub struct DelaunayTriangulation<V, K, L = DelaunayTreeLocate<<V as HasPosition>::Point>> where
    V: HasPosition2D,
    K: DelaunayKernel<<V::Point as PointN>::Scalar>,
    V::Point: TwoDimensional,
    L: DelaunayLocateStructure<V::Point>,  { /* fields omitted */ }

A two dimensional Delaunay triangulation.

A Delaunay triangulation is a special triangulation of a set of points that fulfills some suitable properties for geometric operations like interpolation. There is also an own chapter in the user guide covering spade's triangulation.

Objects that are inserted into the triangulation have to implement the HasPosition2D trait. The trait is implemented for all types that implement TwoDimensional, like Point2 from the cgmath and nalgebra package or [S; 2] for S: SpadeNum.

A straightforward Delaunay triangulation implementation will suffer from precision problems: various geometric queries can fail if imprecise calculations (like native f32 / f64 operations) are used. Those failures can yield to incorrect results or panics at runtime. To prevent those crashes, Spade offers a few "calculation kernels" that may fit the individual needs of an application. See spade::kernels for more information.

Example

extern crate nalgebra;
extern crate spade;

use nalgebra::Point2;
use spade::delaunay::FloatDelaunayTriangulation;

  let mut delaunay = FloatDelaunayTriangulation::with_walk_locate();
  delaunay.insert(Point2::new(0.0, 1.0));
  delaunay.insert(Point2::new(0.0, -1.0));
  delaunay.insert(Point2::new(1.0, 0.0));
  delaunay.insert(Point2::new(-1.0, 0.0));
  delaunay.insert(Point2::new(0.0, 0.0));
  for face in delaunay.triangles() {
    let triangle = face.as_triangle();
    println!("Found triangle: {:?} -> {:?} -> {:?}", *triangle[0], *triangle[1], *triangle[2]);
  }
  for edge in delaunay.edges() {
    println!("Found an edge: {:?} -> {:?}", *edge.from(), *edge.to());
  }

Iterating

A triangulation consists of three elements - vertices, edges and triangles - that can be iterated over. Use vertices() edges() and triangles() to call appropriate, non-mutating iterators.

Infinite face and convex hull

Every triangulation is surrounded by the infinite face. This face can be retrieved by calling infinite_face(). Iterating the adjacent edges of the infinite face will yield the triangulation's convex hull. See FaceHandle::adjacent_edges().

Mutating

Vertices can be added and removed. Mutation of vertex data is possible with lookup_mut(..), although this mutation must not alter the vertex position.

Interpolation

Vertices can store various user data, the points could for example represent meteorological samples measuring temperature or air moisture. These values can be interpolated smoothly within the triangulation's convex hull using a variety of interpolation methods. There are currently four supported interpolation methods:

barycentric_interpolation(..)

nn_interpolation(..)

nn_interpolation_c1_sibson(..)

nn_interpolation_c1_farin(..)

Type parameters

DelaunayTriangulation has three type parameters: V, K and L. V: HasPosition2D defines the triangulation's vertex type. K: DelaunayKernel defines the triangulations calculation kernel. For more information, see spade::kernels. L Defines the locate structure. For more information, see DelaunayLocateStructure.

Performance

Performance of insertion, interpolation and other queries heavily relies on

the locate structure being used

the closeness of subsequent queries

the kernel being used Depending on the use case, it can vary between O(1) to O(sqrt(n)) on average. The guide has an own chapter about performance.

Auto Hinting

Since version 1.1, spade uses the result of the last query as hint for the next query when using DelaunayWalkLocate as locate strategy. As a consequence, subsequent queries - like insertion, interpolation or nearest neighbor queries - will run in O(1) if the query locations are close to each other.

Methods

impl<V, K> DelaunayTriangulation<V, K> where
V: HasPosition2D,
K: DelaunayKernel<<V::Point as PointN>::Scalar>,
V::Point: TwoDimensional,

pub fn with_tree_locate() -> DelaunayTriangulation<V, K>

Shorthand constructor for a Delaunay triangulation that is backed up by an r-tree for log(n) insertion and locate time on average.

pub fn with_walk_locate() -> DelaunayTriangulation<V, K, DelaunayWalkLocate>

Shorthand constructor for a Delaunay triangulation that uses the DelaunayWalkLocate strategy for insertion and point location queries. This yields O(sqrt(n)) insertion time on average for randomly generated vertices.

impl<V, K, L> DelaunayTriangulation<V, K, L> where
V: HasPosition2D,
K: DelaunayKernel<<V::Point as PointN>::Scalar>,
V::Point: TwoDimensional,
L: DelaunayLocateStructure<V::Point>,

pub fn new() -> DelaunayTriangulation<V, K, L>

Creates a new Delaunay triangulation.

Using this method directly can be a bit cumbersome due to type annotations, consider using The short hand definitions FloatDelaunayTriangulation or IntDelaunayTriangulation in combination with the methods with_walk_locate and with_tree_locate.

Example

use spade::delaunay::{DelaunayTriangulation, DelaunayWalkLocate};
use spade::kernels::FloatKernel;
let mut triangulation = DelaunayTriangulation::<_, FloatKernel, DelaunayWalkLocate>::new();

Usually, the omitted types (the triangulation's vertex type) can be inferred from a call to insert.

pub fn vertex(&self, handle: FixedVertexHandle) -> VertexHandle<V>

Creates a dynamic vertex handle from a fixed vertex handle.

May panic if the handle was invalidated by a previous vertex removal.

pub fn vertex_mut(&mut self, handle: FixedVertexHandle) -> &mut V

Returns a mutable reference to the vertex data referenced by a FixedVertexHandle.

pub fn face(&self, handle: FixedFaceHandle) -> FaceHandle<V>

Creates a dynamic face handle from a fixed face handle.

May panic if the faces was invalidated by a previous vertex removal.

pub fn edge(&self, handle: FixedEdgeHandle) -> EdgeHandle<V>

Creates a dynamic edge handle from a fixed edge handle.

May panic if the handle was invalidated by a previous vertex removal.

pub fn num_vertices(&self) -> usize

Returns the number of vertices in this triangulation.

pub fn num_faces(&self) -> usize

Returns the number of faces in this triangulation.

This count does include the infinite face.

pub fn num_triangles(&self) -> usize

Returns the number of triangles in this triangulation.

As there is always exactly one face not being a triangle, this is equivalent to self.num_faces() - 1.

pub fn num_edges(&self) -> usize

Returns the number of edges in this triangulation.

pub fn triangles(&self) -> FacesIterator<V>

Returns an iterator over all triangles.

pub fn edges(&self) -> EdgesIterator<V>

Returns an iterator over all undirected edges.

pub fn vertices(&self) -> VerticesIterator<V>

Returns an iterator over all vertices.

pub fn infinite_face(&self) -> FaceHandle<V>

Returns a handle to the infinite face.

pub fn is_degenerate(&self) -> bool

Returns true if the triangulation is degenerate

A triangulation is degenerate if all vertices of the triangulation lie on one line.

pub fn get_edge_from_neighbors(
&self,
from: FixedVertexHandle,
to: FixedVertexHandle
) -> Option<EdgeHandle<V>>

Returns an edge between two vertices.

If the edge does not exist, None is returned. This operation runs in O(n) time, where n is the degree of from.

pub fn nearest_neighbor(&self, point: &V::Point) -> Option<VertexHandle<V>>

Locates the nearest neighbor for a given point.

Returns None if the triangulation is empty.

pub fn locate(
&self,
point: &V::Point
) -> PositionInTriangulation<VertexHandle<V>, FaceHandle<V>, EdgeHandle<V>>

Returns information about the location of a point in a triangulation.

pub fn locate_vertex(&self, point: &V::Point) -> Option<VertexHandle<V>>

Locates a vertex at a given position.

Returns None if the point could not be found.

pub fn locate_with_hint(
&self,
point: &V::Point,
hint: FixedVertexHandle
) -> PositionInTriangulation<VertexHandle<V>, FaceHandle<V>, EdgeHandle<V>>

Returns information about the location of a point in a triangulation.

Additionally, a hint can be given to speed up computation. The hint should be a vertex close to the position that is being looked up.

pub fn insert(&mut self, t: V) -> FixedVertexHandle

Inserts a new vertex into the triangulation.

This operation runs in O(log(n)) on average when using a tree lookup to back up the triangulation, or in O(sqrt(n)) when using a walk lookup. n denotes the number of vertices, the given running times assume that input data is given uniformly randomly distributed. If the point has already been contained in the triangulation, the old vertex is overwritten.

Returns a handle to the new vertex. Use this handle with DelaunayTriangulation::vertex(..) to refer to it.

pub fn insert_with_hint(
&mut self,
t: V,
hint: FixedVertexHandle
) -> FixedVertexHandle

Inserts a new vertex into the triangulation.

A hint can be given to speed up the process. The hint should be a handle of a vertex close to the new vertex. This method is recommended in combination with DelaunayWalkLocate, in this case the insertion time can be reduced to O(1) on average if the hint is close. If the hint is randomized, running time will be O(sqrt(n)) on average with an O(n) worst case.

pub fn locate_and_remove(&mut self, point: &V::Point) -> Option<V>

Attempts to remove a vertex from the triangulation.

Returns the removed vertex data if it could be found.

Handle invalidation

This method will invalidate all vertex, edge and face handles upon successful removal.

pub fn remove(&mut self, vertex: FixedVertexHandle) -> V

Removes a vertex from the triangulation.

This operation runs in O(n²), where n is the degree of the removed vertex.

Handle invalidation

This method will invalidate all vertex, edge and face handles.

impl<V, K> DelaunayTriangulation<V, K, DelaunayTreeLocate<V::Point>> where
V: HasPosition2D,
K: DelaunayKernel<<V::Point as PointN>::Scalar>,
V::Point: TwoDimensional,

pub fn lookup(&self, point: &V::Point) -> Option<VertexHandle<V>>

Deprecated since 1.3.0
: Use locate_vertex instead

Checks if the triangulation contains an object with a given coordinate.

pub fn lookup_and_remove(&mut self, point: &V::Point) -> Option<V>

Deprecated since 1.3
: Use locate_and_remove instead

Removes a vertex at a given position from the triangulation.

This operation runs in O(n²), where n is the degree of the removed vertex.

Handle invalidation

This method will invalidate all vertex, edge and face handles upon successful removal.

impl<V, K, L> DelaunayTriangulation<V, K, L> where
V: HasPosition2D,
<V::Point as PointN>::Scalar: SpadeFloat,
K: DelaunayKernel<<V::Point as PointN>::Scalar>,
L: DelaunayLocateStructure<V::Point>,
V::Point: TwoDimensional,

pub fn barycentric_interpolation<F>(
&self,
point: &V::Point,
f: F
) -> Option<<V::Point as PointN>::Scalar> where
F: Fn(&V) -> <V::Point as PointN>::Scalar,

Performs a barycentric interpolation. Returns None if the triangulation has no triangles yet. Points outside of the convex hull will be interpolated as well. The other interpolation methods are used very similarly, check their documentation for an example.

pub fn nn_interpolation<F>(
&self,
point: &V::Point,
f: F
) -> Option<<V::Point as PointN>::Scalar> where
F: Fn(&V) -> <V::Point as PointN>::Scalar,

Performs a natural neighbor interpolation for a given position.

Returns None if the triangulation has no triangles yet. Points outside of the convex hull will be interpolated as well.

Example

use spade::HasPosition;
use spade::delaunay::{FloatDelaunayTriangulation};

struct PointWithHeight {
  point: Point2<f64>,
  height: f64,
}

impl HasPosition for PointWithHeight {
  type Point = Point2<f64>;
    fn position(&self) -> Point2<f64> {
      self.point
    }
}

fn main() {
  let mut delaunay = FloatDelaunayTriangulation::with_walk_locate();
  delaunay.insert(PointWithHeight { point: Point2::new(0.0, 0.0), height: 5. });
  delaunay.insert(PointWithHeight { point: Point2::new(1.0, 0.0), height: 0. });
  delaunay.insert(PointWithHeight { point: Point2::new(0.0, 1.0), height: 0. });
  let lookup = Point2::new(0.2, 0.2);
  // Interpolate the points height
  let interpolated = delaunay.nn_interpolation(&lookup, |p| p.height).unwrap();
  // and insert it afterwards.
  delaunay.insert(PointWithHeight { point: lookup, height: interpolated });
  // Data points themselves will always yield their own height
  assert_eq!(delaunay.nn_interpolation(&Point2::new(0.0, 0.0), |p| p.height),
             Some(5.0));
}

pub fn estimate_normal<F, RV>(&self, v: FixedVertexHandle, f: &F) -> RV where
F: Fn(&V) -> <V::Point as PointN>::Scalar,
RV: ThreeDimensional<Scalar = <V::Point as PointN>::Scalar>,

Estimates a normal value for a given vertex.

This assumes that the triangulation models some kind of height field, given by the function f. The normal is the weighted and normalized average of the normals of all triangles adjacent to the given vertex.

pub fn estimate_gradient<F>(&self, v: FixedVertexHandle, f: &F) -> V::Point where
F: Fn(&V) -> <V::Point as PointN>::Scalar,

Estimates and returns the gradient for a single vertex in this triangulation.

pub fn nn_interpolation_c1_sibson<F, G>(
&self,
point: &V::Point,
flatness: <V::Point as PointN>::Scalar,
f: F,
g: G
) -> Option<<V::Point as PointN>::Scalar> where
F: Fn(&V) -> <V::Point as PointN>::Scalar,
G: Fn(&Self, &VertexHandle<V>) -> V::Point,

Interpolates a data point on this triangulation according to Sibson's c1 interpolant.

The interpolation given by nn_interpolation is not differentiable at the triangulation's data points. Sibson introduced another interpolation scheme that takes the gradient of each data point into account and offers an interpolation that is differentiable (c1) at the data points. The interpolation needs to know the gradients of the points natural neighbors, though. Spade can estimate them automatically, see estimate_gradient and estimate_gradients. The value that should be interpolated is given by f, the gradient of a vertex must be given by g.

flatness

An additional flatness factor determines how flat the triangulation will be around the datapoints. A flatness factor of 0.5 is the factor used in sibson's original interpolant. A flatness of 0.0 or lower is nearly identical to sibson's original interpolant (nn_interpolation(..)). A factor of (exactly) 1.0 should yield best performance since an exponentiation can be omitted.

Example

 
use spade::delaunay::FloatDelaunayTriangulation;
use spade::HasPosition;

struct PointWithHeight {
  point: Point2<f64>,
  gradient: Point2<f64>,
  height: f64,
}

impl HasPosition for PointWithHeight {
  type Point = Point2<f64>;
    fn position(&self) -> Point2<f64> {
      self.point
    }
}

impl PointWithHeight {
  fn new(point: Point2<f64>, height: f64) -> PointWithHeight {
    // Initialize the gradient to any value since it will be overwritten
    PointWithHeight { point: point, height: height, gradient: Point2::new(0., 0.) }
  }
}

fn main() {
  let mut delaunay = FloatDelaunayTriangulation::with_walk_locate();
  // Insert some points here... (skipped)
  // Estimate all gradients and store them:
  delaunay.estimate_gradients(&(|v: &PointWithHeight| v.height),
                              &(|v: &mut PointWithHeight, g| v.gradient = g));
   
  // Now we can use the gradients for interpolation, flatness is set to 2.0:
  let interpolated = delaunay.nn_interpolation_c1_sibson(
     &Point2::new(0.5, 0.2), 2.0, |v| v.height, |_, v| v.gradient);
  println!("interpolated: {}", interpolated.unwrap());
}

pub fn nn_interpolation_c1_farin<F, G>(
&self,
point: &V::Point,
f: F,
g: G
) -> Option<<V::Point as PointN>::Scalar> where
F: Fn(&V) -> <V::Point as PointN>::Scalar,
G: Fn(&Self, &VertexHandle<V>) -> V::Point,

Interpolates a data point on this triangulation using Farin's c1 interpolant.

This method is used in the same way as nn_interpolation_c1_sibson. The resulting interpolant is very similar to sibson's c1 interpolant but uses a different algorithm.

impl<V, K, L> DelaunayTriangulation<V, K, L> where
V: HasPosition2D,
<V::Point as PointN>::Scalar: SpadeFloat,
K: DelaunayKernel<<V::Point as PointN>::Scalar>,
L: DelaunayLocateStructure<V::Point>,
V::Point: TwoDimensional,

pub fn estimate_normals<F, G, RV>(&mut self, f: &F, g: G) where
F: Fn(&V) -> <V::Point as PointN>::Scalar,
G: Fn(&mut V, RV),
RV: ThreeDimensional<Scalar = <V::Point as PointN>::Scalar>,

Estimates a normal for each vertex in the triangulation.

f must yield a 'height' value for each vertex, g is a callback function that can be used to store the calculated normals.

 
use spade::delaunay::FloatDelaunayTriangulation;
use spade::HasPosition;

struct PointWithHeight {
  point: Point2<f64>,
  normal: Point3<f64>,
  height: f64,
}

impl HasPosition for PointWithHeight {
  type Point = Point2<f64>;
    fn position(&self) -> Point2<f64> {
      self.point
    }
}
impl PointWithHeight {
  fn new(point: Point2<f64>, height: f64) -> PointWithHeight {
    PointWithHeight { point: point, height: height, normal: Point3::new(0., 0., 0.) }
  }
}

fn main() {
  let mut delaunay = FloatDelaunayTriangulation::with_walk_locate();
  // Insert some points here... (skipped)
  // Then, estimate all normals at once:
  delaunay.estimate_normals(&(|v: &PointWithHeight| v.height), &(|v: &mut PointWithHeight, n| v.normal = n));
   
  // And print them
  for vertex in delaunay.vertices() {
     println!("vertex: {:?}, normal: {:?}", vertex.position(), vertex.normal);
  }
}

pub fn estimate_gradients<F, G>(&mut self, f: &F, g: &G) where
F: Fn(&V) -> <V::Point as PointN>::Scalar,
G: Fn(&mut V, V::Point),

Estimates gradients for all vertices in this triangulation.

f yields the value for which a gradient should be calculated, g can be used to store the gradient. See estimate_normals for a similar example.

Internally, the normal for each vertex is calculated first. Then, an appropriate gradient is calculated.