Struct spade::delaunay::EdgeHandle

pub struct EdgeHandle<'a, V> where
    V: 'a,  { /* fields omitted */ }

A handle to a directed edge.

This handle has methods that can be used to retrieve the edges neighboring edges, faces and vertices.

Methods

impl<'a, V> EdgeHandle<'a, V> where
V: 'a,

fn fix(&self) -> FixedEdgeHandle

Creates a fixed vertex handle from this dynamic handle.

fn from(&self) -> VertexHandle<'a, V>

Returns the edge's source vertex.

fn o_next(&self) -> EdgeHandle<'a, V>

Returns the oriented next edge.

The oriented next edge shares the same face as this edge. When traversing the face's edges in oriented order, this edge is the predecessor of the oriented next edge. "Oriented" means counterclockwise for right handed coordinate systems.

fn o_prev(&self) -> EdgeHandle<'a, V>

Returns the oriented previous edge.

The oriented previous edge shares the same face as this edge. When traversing the face's edges in oriented order, this edge is the successor of the oriented previous edge. "Oriented" means counterclockwise for right handed coordinate systems.

fn o_next_iterator(&self) -> ONextIterator<'a, V>

Returns an iterator over all edges sharing the same face as this edge.

The face's edges will be traversed in oriented order. This order is counterclockwise for right handed coordinate systems or clockwise for left handed systems.

fn to(&self) -> VertexHandle<'a, V>

Returns the edges destination vertex.

fn face(&self) -> FaceHandle<'a, V>

Returns the face located to the left of this edge.

fn sym(&self) -> EdgeHandle<'a, V>

Returns this edge's mirror edge.

fn cw(&self) -> EdgeHandle<'a, V>

Returns the next edge in clockwise direction.

Note that this assumes that you use a right handed coordinate system, otherwise the sense of orientation is inverted.

fn ccw(&self) -> EdgeHandle<'a, V>

Returns the next edge in counter clockwise direction.

Note that this assumes that you use a right handed coordinate system, otherwise the sense of orientation is inverted.

fn ccw_iter(&self) -> CCWIterator<'a, V>

Returns an iterator over all edges in counter clockwise order.

Note that this assumes that you use a right handed coordinate system, otherwise the sense of orientation is inverted.