Struct spade::delaunay::FaceHandle

pub struct FaceHandle<'a, V, E = ()> where
    V: 'a,
    E: 'a,  { /* fields omitted */ }

A handle to a face.

Used to retrieve its adjacent edges.

Methods

impl<'a, V, E> FaceHandle<'a, V, E> where
V: 'a,
E: Default + 'a,

pub fn as_triangle(&self) -> [VertexHandle<'a, V, E>; 3]

Tries to interpret this face as a triangle, returning its 3 vertices.

Panic

This method will panic if the face does not form a triangle.

pub fn adjacent_edge(&self) -> Option<EdgeHandle<'a, V, E>>

Returns an edge that is adjacent to this face.

If this face has multiple adjacent edges, any of them is returned.

Important traits for ONextIterator<'a, V, E>
impl<'a, V, E> Iterator for ONextIterator<'a, V, E> where
V: 'a,
E: Default + 'a, type Item = EdgeHandle<'a, V, E>;

pub fn adjacent_edges(&self) -> ONextIterator<'a, V, E>

Returns an iterator that iterates over all adjacent edges.

The edges are traversed in oriented order. This order will be counterclockwise for right handed coordinate system or clockwise for left handed systems.

pub fn fix(&self) -> FixedFaceHandle

Creates a fixed face handle from this dynamic face handle.