Type Definition spade::handles::FixedDirectedEdgeHandle

pub type FixedDirectedEdgeHandle = FixedHandleImpl<DirectedEdgeTag, InnerTag>;

Fixed handle to a directed edge.

See also the handles module.

Implementations

impl FixedDirectedEdgeHandle

pub fn rev(self) -> Self

Returns this edge with its direction reversed.

If this edge points from v0 to v1, the returned edge would point from v1 to v0. Calling rev twice will always return the original vertex.

pub fn as_undirected(self) -> FixedUndirectedEdgeHandle

Converts this directed edge handle into an undirected edge handle.

See also the handles module for more information.