Type Definition spade::handles::FixedFaceHandle

pub type FixedFaceHandle<InnerOuter> = FixedHandleImpl<FaceTag, InnerOuter>;

“Fixed handle to a face.

The type parameter is either InnerTag or PossiblyOuterTag, depending on the face type.

“See also the handles module

Implementations

impl FixedFaceHandle<PossiblyOuterTag>

pub fn is_outer(&self) -> bool

Returns true if this face is the single outer face.

pub fn as_inner(&self) -> Option<FixedFaceHandle<InnerTag>>

Converts this face handle to an inner face.

Returns None if this handle refers to the single outer face.