Type Definition spade::handles::FixedUndirectedEdgeHandle

pub type FixedUndirectedEdgeHandle = FixedHandleImpl<UndirectedEdgeTag, InnerTag>;

Fixed handle to an undirected edge.

See also the handles module.

Implementations

impl FixedUndirectedEdgeHandle

pub fn as_directed(&self) -> FixedDirectedEdgeHandle

Converts this directed edge into an undirected edge handle.

Any of the two directed edges may be returned.

See also FixedDirectedEdgeHandle::as_undirected