Type Alias spade::handles::FixedVertexHandle

pub type FixedVertexHandle = FixedHandleImpl<VertexTag, InnerTag>;

Fixed handle to a vertex.

See also the handles module.

Aliased Type

struct FixedVertexHandle { /* private fields */ }

Implementations

impl FixedVertexHandle

pub fn from_index(index: usize) -> Self

Creates a new vertex handle from a usize.

Ideally, this method should only be used in advanced scenarios as it allows to create invalid vertex handles. When possible, attempt to retrieve handles by other means instead (see crate::handles).

Panics

Panics if index >= 2^32