Type Alias spade::handles::UndirectedVoronoiEdge

pub type UndirectedVoronoiEdge<'a, V, DE, UE, F> = DynamicHandleImpl<'a, V, DE, UE, F, UndirectedVoronoiEdgeTag, InnerTag>;

A handle to an undirected edge of the Voronoi diagram.

Aliased Type

struct UndirectedVoronoiEdge<'a, V, DE, UE, F> { /* private fields */ }

Implementations

impl<'a, V, DE, UE, F> UndirectedVoronoiEdge<'a, V, DE, UE, F>

pub fn vertices(&self) -> [VoronoiVertex<'a, V, DE, UE, F>; 2]

Returns the edge’s two vertices.

The vertices are returned in any order.

pub fn as_directed(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Converts this undirected handle into a directed edge handle.

pub fn as_delaunay_edge(&self) -> UndirectedEdgeHandle<'a, V, DE, UE, F>

Returns the dual edge of the Delaunay triangulation.