Type Definition spade::handles::VoronoiFace

pub type VoronoiFace<'a, V, DE, UE, F> = DynamicHandleImpl<'a, V, DE, UE, F, VoronoiFaceTag, PossiblyOuterTag>;

A handle to a face of the voronoi diagram.

Implementations

impl<'a, V, DE, UE, F> VoronoiFace<'a, V, DE, UE, F>

pub fn as_delaunay_vertex(&self) -> VertexHandle<'a, V, DE, UE, F>

Converts this face into its dual vertex of the Delaunay Triangulation.

pub fn adjacent_edges(
&self
) -> impl DoubleEndedIterator<Item = DirectedVoronoiEdge<'a, V, DE, UE, F>>

Returns an iterator that returns all edges adjacent to this face.

The edges are returned in clockwise order.