Struct spade::internals::DynamicHandleImpl

pub struct DynamicHandleImpl<'a, V, DE, UE, F, Type, InnerOuter: InnerOuterMarker> { /* private fields */ }

Internal type definition that is only exposed for documentation purposes.

Rust will currently not generate documentation for type definitions depending pub(crate) types, see #32077.

Do not use these types. Their removal from the public API will not be considered a breaking change.

Refer to the handles module for the handle types that should be used instead.

Implementations

impl<'a, V, DE, UE, F, Type: Copy, InnerOuter: InnerOuterMarker> DynamicHandleImpl<'a, V, DE, UE, F, Type, InnerOuter>

pub fn fix(&self) -> FixedHandleImpl<Type, InnerOuter>

Converts this dynamic handle to its fixed variant.

See also the handles module

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, DirectedEdgeTag, InnerTag>

pub fn vertices(&self) -> [VertexHandle<'a, V, DE, UE, F>; 2]

Returns the edge’s two vertices.

The first vertex is self.from(), the second vertex is self.to().

pub fn from(&self) -> VertexHandle<'a, V, DE, UE, F>

Returns the edge’s origin vertex.

pub fn to(&self) -> VertexHandle<'a, V, DE, UE, F>

Returns the edge’s destination vertex.

pub fn rev(&self) -> Self

Returns this edge in reversed direction.

pub fn opposite_vertex(&self) -> Option<VertexHandle<'a, V, DE, UE, F>>

Returns the vertex which lies opposite of this edge.

This is equal to calling self.prev().from() or self.next().to(). Returns None if this edge is part of the convex hull.

pub fn next(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns the oriented next edge.

The oriented next edge shares the same face as this edge. When traversing the face’s edges in oriented order, this edge is the predecessor of the oriented next edge. “Oriented” means counterclockwise for right handed coordinate systems.

pub fn prev(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns the oriented previous edge.

The oriented previous edge shares the same face as this edge. When traversing the face’s edges in oriented order, this edge is the successor of the oriented previous edge. “Oriented” means counterclockwise for right handed coordinate systems.

pub fn face(&self) -> FaceHandle<'a, PossiblyOuterTag, V, DE, UE, F>

Returns the face located to the left of this edge.

pub fn cw(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns the next edge in clockwise direction.

Note that this assumes that you use a right handed coordinate system, otherwise the sense of orientation is inverted.

pub fn ccw(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns the next edge in counter clockwise direction.

Note that this assumes that you use a right handed coordinate system, otherwise the sense of orientation is inverted.

pub fn data(&self) -> &'a DE

Returns a reference to the data associated with this directed edge.

Use Triangulation::directed_edge_data_mut to modify the edge data.

pub fn as_undirected(self) -> UndirectedEdgeHandle<'a, V, DE, UE, F>

Converts this directed edge handle into an undirected edge handle.

See also the handles module.

pub fn is_outer_edge(&self) -> bool

Returns true if this edge is adjacent to the outer face.

pub fn is_part_of_convex_hull(&self) -> bool

Returns true if either this edge or its reversed edge is adjacent to the outer face.

pub fn as_voronoi_edge(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Converts this edge into its dual voronoi edge.

See also as_delaunay_edge.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, DirectedEdgeTag, InnerTag> where
V: HasPosition,

pub fn opposite_position(&self) -> Option<Point2<V::Scalar>>

Returns the position of the vertex opposite of this edge.

See also opposite_vertex(). Returns None if this edge is an outer edge.

pub fn length_2(&self) -> V::Scalar

Returns the squared length of this edge.

pub fn side_query(&self, query_point: Point2<V::Scalar>) -> LineSideInfo

Identifies on which side of this edge a point lies.

pub fn project_point(
&self,
query_point: Point2<V::Scalar>
) -> PointProjection<V::Scalar>

Indicates the position of a point being projected onto this edge.

A point’s projection can either come before, on or behind this edge. Note that this method may return inaccurate results due to rounding issues.

An image displaying differently colored areas which would result in different point projections

Example

use spade::{Point2, Triangulation, DelaunayTriangulation};

let from = Point2::new(0.0, 0.0);
let to = Point2::new(2.0, 0.0);

let mut triangulation: DelaunayTriangulation<_> = Default::default();
let v0 = triangulation.insert(from)?;
let v1 = triangulation.insert(to)?;
// This edge goes from "from" to "to"
let edge = triangulation.get_edge_from_neighbors(v0, v1).unwrap();

// These vertices are all projected before the edge
assert!(edge.project_point(Point2::new(-0.2, 0.0)).is_before_edge());
assert!(edge.project_point(Point2::new(-1002.0, -12.0)).is_before_edge());

// These vertices are all projected onto the edge
assert!(edge.project_point(Point2::new(1.0, 5.0)).is_on_edge());
assert!(edge.project_point(Point2::new(0.5, -2.0)).is_on_edge());
assert!(edge.project_point(Point2::new(1.0, 0.0)).is_on_edge());

// These vertices are all projected behind the edge
assert!(edge.project_point(Point2::new(4.000001, 0.0)).is_behind_edge());
assert!(edge.project_point(Point2::new(5.0, -12.0)).is_behind_edge());

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, UndirectedVoronoiEdgeTag, InnerTag>

pub fn vertices(&self) -> [VoronoiVertex<'a, V, DE, UE, F>; 2]

Returns the edge’s two vertices.

The vertices are returned in any order.

pub fn as_directed(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Converts this undirected handle into a directed edge handle.

pub fn as_delaunay_edge(&self) -> UndirectedEdgeHandle<'a, V, DE, UE, F>

Returns the dual edge of the Delaunay triangulation.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, UndirectedEdgeTag, InnerTag>

pub fn vertices(&self) -> [VertexHandle<'a, V, DE, UE, F>; 2]

Returns the edge’s two vertices.

The vertices are returned in any order.

pub fn as_directed(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Converts this directed edge into an undirected edge handle.

pub fn as_voronoi_edge(&self) -> UndirectedVoronoiEdge<'a, V, DE, UE, F>

Returns the dual edge in the Voronoi diagram.

pub fn data(&self) -> &UE

Returns a reference to the data associated with this directed edge.

Use Triangulation::undirected_edge_data_mut to modify the edge data.

pub fn is_part_of_convex_hull(&self) -> bool

Returns true if the outer face is adjacent to any side of this undirected edge.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, UndirectedEdgeTag, InnerTag> where
V: HasPosition,

pub fn positions(&self) -> [Point2<V::Scalar>; 2]

Returns the end positions of this edge.

The positions are returned in any order.

pub fn length_2(&self) -> V::Scalar

Returns the squared length of this edge

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, UndirectedEdgeTag, InnerTag> where
V: HasPosition,
V::Scalar: Float,

pub fn distance_2(&self, query_point: Point2<V::Scalar>) -> V::Scalar

Returns the squared distance of a point to this edge.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, FaceTag, InnerOuter>

pub fn adjacent_edges(&self) -> [DirectedEdgeHandle<'a, V, DE, UE, F>; 3]

Returns the three inner edges adjacent to this face.

The edges are returned in counter clockwise order.

pub fn adjacent_edge(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns an edge that is adjacent to this face.

If this face has multiple adjacent edges, any of them is returned.

pub fn vertices(&self) -> [VertexHandle<'a, V, DE, UE, F>; 3]

Returns the face’s three vertices.

The vertices are returned in counter clockwise order.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, FaceTag, InnerOuter> where
V: HasPosition,

pub fn positions(&self) -> [Point2<V::Scalar>; 3]

Returns the positions of the face’s vertices

The positions are returned in counter clockwise order.

pub fn area(&self) -> V::Scalar

Returns the triangle’s area.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, FaceTag, InnerOuter> where
V: HasPosition,
V::Scalar: Float,

pub fn distance_2(&self, query_point: Point2<V::Scalar>) -> V::Scalar

Returns the squared distance of a point to this triangle.

The distance of a point inside the triangle is zero.

pub fn center(&self) -> Point2<V::Scalar>

Returns the face’s center point.

The center point is the average position of its vertices.

pub fn circumcircle(&self) -> (Point2<V::Scalar>, V::Scalar)

Returns the face’s circumcircle center and the squared radius of the circumcircle.

The circumcircle is the unique circle that intersects all three vertices of the face.

pub fn circumcenter(&self) -> Point2<V::Scalar>

Returns the face’s circumcenter.

The circumcenter is the center of the circumcircle.

pub fn barycentric_interpolation(
&self,
coordinate: Point2<V::Scalar>
) -> [V::Scalar; 3]

Returns the barycentric coordinates of a point relative to this face.

The returned coordinates will sum up to 1.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, VertexTag, InnerTag> where
V: HasPosition,

pub fn position(&self) -> Point2<V::Scalar>

Returns the position of this vertex.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, VertexTag, InnerTag>

pub fn out_edges(
&self
) -> CircularIterator<'a, V, DE, UE, F, CCWEdgesNextBackFn>

Returns all directed edges going out of this vertex.

The edges are returned in counter clockwise order, beginning at an arbitrary edge.

A possible iteration order of v.out_edges()

Note: The returned iterator implements DoubleEndedIterator, allowing traversal in clockwise order.

pub fn out_edge(&self) -> Option<DirectedEdgeHandle<'a, V, DE, UE, F>>

Returns an outgoing edge of this vertex.

If the vertex has multiple outgoing edges, any of them is returned.

pub fn data(&self) -> &V

Returns the data associated with this vertex.

pub fn as_voronoi_face(&self) -> VoronoiFace<'a, V, DE, UE, F>

Returns the voronoi face that corresponds to this vertex of the Delaunay triangulation.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, DirectedEdgeTag, InnerTag> where
V: HasPosition,
V::Scalar: Float,

pub fn distance_2(&self, query_point: Point2<V::Scalar>) -> V::Scalar

Returns the squared distance of a point to this edge.

pub fn nearest_point(&self, query_point: Point2<V::Scalar>) -> Point2<V::Scalar>

Yields the nearest point on this edge.

impl<'a, V, DE, UE, F, InnerOuter: InnerOuterMarker> DynamicHandleImpl<'a, V, DE, UE, F, FaceTag, InnerOuter>

pub fn data(&self) -> &F

Returns a reference to the data associated with this face.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, FaceTag, InnerOuter>

pub fn is_outer(&self) -> bool

Returns true if this handle refers to the single outer face.

pub fn as_inner(&self) -> Option<FaceHandle<'a, InnerTag, V, DE, UE, F>>

Converts this possibly outer face handle into an inner face handle.

Returns None if this handle refers to the outer face.

pub fn adjacent_edge(&self) -> Option<DirectedEdgeHandle<'a, V, DE, UE, F>>

Returns an edge that is adjacent to this face.

The returned edge has this face on its left side. Returns None if the triangulation has only one or no vertices.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, VoronoiFaceTag, PossiblyOuterTag>

pub fn as_delaunay_vertex(&self) -> VertexHandle<'a, V, DE, UE, F>

Converts this face into its dual vertex of the Delaunay Triangulation.

pub fn adjacent_edges(
&self
) -> impl DoubleEndedIterator<Item = DirectedVoronoiEdge<'a, V, DE, UE, F>>

Returns an iterator that returns all edges adjacent to this face.

The edges are returned in clockwise order.

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, DirectedVoronoiEdgeTag, InnerTag>

pub fn to(&self) -> VoronoiVertex<'a, V, DE, UE, F>

Returns the voronoi edge’s destination.

pub fn from(&self) -> VoronoiVertex<'a, V, DE, UE, F>

Returns the voronoi vertex from which this edge originates.

pub fn face(&self) -> VoronoiFace<'a, V, DE, UE, F>

Returns the Voronoi face to the left of this Voronoi edge

pub fn as_undirected(&self) -> UndirectedVoronoiEdge<'a, V, DE, UE, F>

Converts this directed edge handle into an undirected edge handle.

See also the handles module for more information.

pub fn as_delaunay_edge(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns the directed dual edge of the underlying Delaunay triangulation.

The dual edge is always orthogonal to to this edge.

Shows various inner voronoi edges (blue) and their dual Delaunay edges (orange)

pub fn rev(&self) -> Self

Returns this edge with its direction reversed.

pub fn next(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Returns the edge that is connected to this edge in counter clockwise order.

See also prev

pub fn prev(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Returns the edge that is connected to this edge in clockwise order.

See also next

impl<'a, V, DE, UE, F> DynamicHandleImpl<'a, V, DE, UE, F, DirectedVoronoiEdgeTag, InnerTag> where
V: HasPosition,

pub fn direction_vector(&self) -> Point2<V::Scalar>

Returns a vector that is parallel to the voronoi edge.

This vector is obtained by rotating the dual Delaunay edge by 90° degree The returned vector is not necessarily normalized.