Struct spade::kernels::TrivialKernel

pub struct TrivialKernel {}

Offers fast and possibly inaccurate geometric calculations.

Use this kernel if you are working with small integral coordinates (e.g. Point2<i64> in the range ±100000) as coordinate vector. Offers best performance. Run in debug mode to be notified if an overflow occurs. Using this kernel with f32 or f64 coordinates can lead to runtime panics, incorrect triangulations or infinite loops and is not recommended.

If your application runs into over / underflow issues, consider using AdaptiveIntKernel.