Struct spade::primitives::EdgeSideInfo

pub struct EdgeSideInfo<S> {
    // some fields omitted
}

Yields information about on which side of a line a point lies.

Methods

impl<S> EdgeSideInfo<S> where S: SpadeNum

fn is_on_left_side(&self) -> bool

Returns true if the query point lies on the left side of the query edge.

fn is_on_right_side(&self) -> bool

Returns true if the query point lies on the right side of the query edge.

fn is_on_left_side_or_on_line(&self) -> bool

Returns true if the query point lies on the left side or on the query edge.

fn is_on_right_side_or_on_line(&self) -> bool

Returns true if the query point lies on the right side or on the query edge.

fn is_on_line(&self) -> bool

Returns true if the query point lies on an infinite line going through the edge's start and end point.