Struct spade::primitives::SimpleEdge

pub struct SimpleEdge<V: VectorN> {
    pub from: V,
    pub to: V,
}

An edge defined by it's two end points.

Fields
from: V

The edge's origin.

to: V

The edge's destination.

Methods

impl<V> SimpleEdge<V> where V: VectorN

fn new(from: V, to: V) -> SimpleEdge<V>

Creates a new edge from from to to.

fn is_projection_on_edge(&self, query_point: &V) -> bool

Projects a point onto the infinite line going through the edge's start and end point and returns true if the projected points lies between from and to.

impl<V> SimpleEdge<V> where V: TwoDimensional

fn side_query(&self, q: &V) -> EdgeSideInfo<V::Scalar>

Determines on which side of this edge a given point lies.

impl<V> SimpleEdge<V> where V: VectorN, V::Scalar: SpadeFloat

fn nearest_point(&self, query_point: &V) -> V

Yields the nearest point on this edge.

fn projection_distance2(&self, query_point: &V) -> V::Scalar

Returns the minimal squared distance of a given point to its projection onto the infinite edge going through this edge's start and end point.

fn project_point(&self, query_point: &V) -> V::Scalar

Projects a point on this line and returns its relative position.

This method will return a value between 0. and 1. (linearly interpolated) if the projected point lies between self.from and self.to, a value close to zero (due to rounding errors) if the projected point is equal to self.from and a value smaller than zero if the projected point lies "before" self.from. Analogously, a value close to 1. or greater than 1. is returned if the projected point is equal to or lies behind self.to.