Struct spade::FloatKernel

pub struct FloatKernel {
}

Offers a fast, precise kernel working with f64 coordinates.

Performing a delaunay triangulation is often a tradeoff between accuracy and speed: a triangulation working on native f64-operations will fail in rare cases due to rounding errors, while switching to precise floats (like BigRationals from the num crate) reduces performance by several orders of magnitude. This kernel works with adaptive precision: if a calculation is inaccurate, it will increase its precision until the result is accurate enough. Since most calculations are accurate enough in their simplest form, only the overhead of checking the precision is usually encountered. For more information, refer to this link describing the techniques behind the adaptive approach.