Struct spade::RTree

pub struct RTree<T> where T: SpatialObject {
    // some fields omitted
}

A rust implementation of n dimensional r*-trees

R-trees provide efficient nearest-neighbor searches for many objects. R*-trees ("R-Star-Trees") are a common variant of r-trees and use more advanced heuristics to improve query performance. This struct implements r*-trees, despite its name. Instead of linear time complexity, r-trees yield logarithmic complexity for look-up operations and nearest neighbor queries. Inserting into an r-tree runs in O(log(n)) time on average. Some simple simple geometric primitives that can be inserted into an r-tree can be found in the primitives module. If your object is not among those, consider implementing the SpatialObject trait.

Note that the rtree-infrastructure works with 2, 3 and 4 dimensional vectors from the nalgebra and cgmath package with both integral and floating point scalar types:

  let mut tree = RTree::new();
  tree.insert(Vector4::new(13i32, 10, 10, 37)); // Note that this is an integer vector

Basic Example

extern crate cgmath; // Alternativeley: use nalgebra
extern crate spade;

use spade::RTree;
use cgmath::Vector2;

fn main() {
let mut rtree = RTree::new();
// Insert two points
rtree.insert(Vector2::new(0.5, 0.5f32));
rtree.insert(Vector2::new(1.0, 1.0f32));

if rtree.lookup(&Vector2::new(0.5, 0.5)).is_some() {
  println!("We'fe found a point at [0.5, 0.5]/");
}
 
let nearest = rtree.nearest_neighbor(&Vector2::new(1.5, 1.5)).unwrap();
println!("nearest neighbor at [1.5, 1.5]: {:?}", nearest);

// Iterate over all elements
for point in rtree.iter() {
  println!("Found point: {:?}", point);
}
}

Methods

impl<T> RTree<T> where T: SpatialObject

fn new() -> RTree<T>

Creates an empty r*-tree.

fn mbr(&self) -> Option<BoundingRect<T::Vector>>

Returns the trees minimal bounding box.

fn size(&self) -> usize

Returns the number of elements contained in this r-tree.

fn iter(&self) -> RTreeIterator<T>

Returns an iterator over all contained elements.

fn insert(&mut self, t: T)

Inserts a new element into the tree.

This will require O(log(n)) operations on average, where n is the number of elements contained in the tree.

fn lookup_and_remove(&mut self, query_point: &T::Vector) -> Option<T>

Searches for an element and removes it.

If the given point is contained by one object in the tree, this object is being removed and returned. If the point is contained by multiple objects, only one of them is removed and returned.

fn lookup(&self, query_point: &T::Vector) -> Option<&T>

Searches for an element at a given position.

If query_point is contained by one object in the tree, this object will be returned. If multiple objects contain the point, only one of them will be returned.

fn lookup_mut(&mut self, query_point: &T::Vector) -> Option<&mut T>

Searches for an element at a given position and returns a mutable reference. If query_point is contained by multiple objects in the tree, one of them will be returned. Do not change the object's minimal bounding box.

fn nearest_neighbor(&self, query_point: &T::Vector) -> Option<&T>

Returns the nearest neighbor.

Returns None if the tree is empty.

fn close_neighbor(&self, query_point: &T::Vector) -> Option<&T>

Returns an object close to a given point. This operation is faster than nearest_neighbor but will not neccessarily yield the real nearest neighbor.

fn nearest_neighbors(&self, query_point: &T::Vector) -> Vec<&T>

Returns the nearest neighbors of a given point.

All returned values will have the exact same distance from the given query point. Returns an empty vector if the tree is empty.

fn nearest_n_neighbors(&self, query_point: &T::Vector, n: usize) -> Vec<&T>

Returns the nearest n neighbors.

fn lookup_in_rectangle(&self, query_rect: &BoundingRect<T::Vector>) -> Vec<&T>

Returns all objects (partially) contained in a rectangle

fn lookup_in_circle(&self, circle_origin: &T::Vector, radius2: &T::Vector::Scalar) -> Vec<&T>

Returns all objects (partially) contained in a circle.

Note that radius2 is the circle's squared radius, not the actual radius. An object is contained if a part of it lies within the circle.

impl<T> RTree<T> where T: SpatialObject + PartialEq

fn remove(&mut self, obj: &T) -> bool

Removes an object from the tree.

Locates and removes an object from the tree, returning true if the element could be removed. If multiple object's are equal to to_remove, only one will be deleted.

fn contains(&self, obj: &T) -> bool

Returns true if a given object is contained in this tree.