Struct spade::TrivialKernel

pub struct TrivialKernel {
}

Offers fast, possibly approximative and possibly inaccurate geometric calculations.

Use this kernel if you are working with small integral coordinates (e.g. Vector2<i64> in the range ±100000) or if you are willing to risk crashes when using Vector2<f64> as coordinate. Offers best performance. Run this in debug mode to be notified early if something goes amiss (e.g. over / underflow, contradictory geometric calculations) since DelaunayTriangulation uses a few debug_asserts to keep its internal structure healthy.

If you run into said problems, consider using FloatKernel or AdaptiveIntKernel.