Struct spade::VertexHandle

pub struct VertexHandle<'a, V, K> where V: HasPosition2D + 'a, K: 'a, V::Vector: TwoDimensional {
    // some fields omitted
}

A reference to a vertex in a triangulation. Note that this handle dereferences to the data it represents.

Example

  let mut delaunay = DelaunayTriangulation::default();
  let fixed_handle = delaunay.insert(Vector2::new(0.0, 1.0));
  delaunay.insert(Vector2::new(0.0, -1.0));
  delaunay.insert(Vector2::new(1.0, 0.0));
  let vertex_handle = delaunay.handle(fixed_handle);
  assert_eq!(vertex_handle.num_neighbors(), 2);
  // This dereferences into a Vector2
  assert_eq!(*vertex_handle, Vector2::new(0.0, 1.0));

Methods

impl<'a, V: HasPosition2D, K> VertexHandle<'a, V, K> where K: DelaunayKernel<V::Vector::Scalar>, V::Vector: TwoDimensional

fn fixed_neighbors(&self) -> &'a [FixedVertexHandle]

Returns the neighbors of this vertex as fixed handles

fn neighbors(&self) -> Box<Iterator<Item=VertexHandle<'a, V, K>> + 'a>

Returns the neighbors an iterator over the neighbors of this vertex.

fn out_edges(&self) -> Box<Iterator<Item=EdgeHandle<'a, V, K>> + 'a>

Returns an iterator over all out edges of this vertex.

fn sector_info(&self, point: &V::Vector) -> SectorInfo

Returns information about the sector in which a given point is contained. A "Sector" is the circle segment between two adjacent out edges.

fn num_neighbors(&self) -> usize

Returns the number of neighbors of this point.

fn fix(&self) -> FixedVertexHandle

Returns this handle as FixedVertexHandle.