Trait spade::DelaunayKernel

pub trait DelaunayKernel<D: SpadeNum>: Sized {
    fn contained_in_circumference<V: TwoDimensional<Scalar=D>>(pa: &V, pb: &V, pc: &V, pd: &V) -> bool { ... }
    fn side_query<V: TwoDimensional<Scalar=D>>(edge: &SimpleEdge<V>, position: &V) -> EdgeSideInfo<D> { ... }
    fn is_ordered_ccw<V: TwoDimensional<Scalar=D>>(v0: &V, v1: &V, v2: &V) -> bool { ... }
    fn new_triangulation<T: HasPosition>() -> DelaunayTriangulation<T, Self> where T::Vector: TwoDimensional<Scalar=D> { ... }
    fn point_on_edge<V: TwoDimensional<Scalar=D>>(edge: &SimpleEdge<V>, position: &V) -> bool { ... }
}

Determines how a delaunay triangulation performs its basic geometry computation.

Every delaunay triangulation is based on two basic geometry operations: orientation tests (on which side of a line lies a point?) and in-circle tests (is a point contained in the circumference of a triangle?). These questions can be answered approximately or precise, their calculation can take overflow issues into account (if handling integer coordinates).

Since each application has different needs, a DelaunayKernel will define how these geometric queries are calculated for each triangulation. It is recommended to use one of the predefined kernels that fits your needs.

Provided Methods

fn contained_in_circumference<V: TwoDimensional<Scalar=D>>(pa: &V, pb: &V, pc: &V, pd: &V) -> bool

Returns true if pd is contained in the circumference of the triangle spanned by pa, pb, pc. pa, pb, pc have to be ordered clockwise.

fn side_query<V: TwoDimensional<Scalar=D>>(edge: &SimpleEdge<V>, position: &V) -> EdgeSideInfo<D>

Returns an EdgeSideInfo yielding on which side of a line a point lies.

fn is_ordered_ccw<V: TwoDimensional<Scalar=D>>(v0: &V, v1: &V, v2: &V) -> bool

Another formulation of side_query, will return true if v0, v1 and v2 are ordered counterclockwise.

fn new_triangulation<T: HasPosition>() -> DelaunayTriangulation<T, Self> where T::Vector: TwoDimensional<Scalar=D>

Creates a new delaunay triangulation using this kernel.

fn point_on_edge<V: TwoDimensional<Scalar=D>>(edge: &SimpleEdge<V>, position: &V) -> bool

Returns if a point lies on the infinite edge going through edge.from and edge.to.