Trait spade::SpadeNum

pub trait SpadeNum: Num + PartialOrd + Into<f64> + From<f32> + Copy + Signed + Debug { }

A coordinate type that can be used with a triangulation.

Internally, most calculations are performed after converting the type into a f64. However, changing this to f32 will reduce the required storage space slightly.

This type should usually be either f32 or f64.