Trait spade::SpatialObject

pub trait SpatialObject {
    type Vector: VectorN;
    fn mbr(&self) -> BoundingRect<Self::Vector>;
    fn distance2(&self, point: &Self::Vector) -> Self::Vector::Scalar;

    fn contains(&self, point: &Self::Vector) -> bool { ... }
}

Describes objects that can be located by r-trees.

See the primitives module for some basic implementations which can also serve as useful examples for own implementations.

Associated Types

type Vector: VectorN

The object's vector type.

Required Methods

fn mbr(&self) -> BoundingRect<Self::Vector>

Returns the object's minimal bounding rectangle.

The minimal bounding rectangle is the smallest axis aligned rectangle that completely contains the object. Note: The rectangle must be as small as possible, otherwise some queries might fail.

fn distance2(&self, point: &Self::Vector) -> Self::Vector::Scalar

Returns the squared euclidean distance from the object's contour. Returns a value samller than zero if the point is contained within the object.

Provided Methods

fn contains(&self, point: &Self::Vector) -> bool

Returns true if a given point is contained in this object.