Trait spade::ThreeDimensional

pub trait ThreeDimensional: PointN {
    fn cross(&self, other: &Self) -> Self { ... }
}

A three dimensional Point.

Some algorithms will only work with three dimensional points, this trait makes sure that only such points can be used.

Provided Methods

fn cross(&self, other: &Self) -> Self

The cross product of this point and another.