Trait spade::TwoDimensional

pub trait TwoDimensional: PointN { }

A two dimensional Point.

Some datastructures will only work if two dimensional points are given, this trait makes sure that only such points can be passed.