Trait spade::VectorN

pub trait VectorN where Self: Clone, Self: Add<Self, Output=Self>, Self: Sub<Self, Output=Self>, Self: Div<Self::Scalar, Output=Self>, Self: Mul<Self::Scalar, Output=Self>, Self: Index<usize, Output=Self::Scalar>, Self: IndexMut<usize, Output=Self::Scalar>, Self: Debug, Self: PartialEq {
    type Scalar: SpadeNum;
    fn from_value(value: Self::Scalar) -> Self;
    fn dimensions() -> usize;

    fn new() -> Self { ... }
    fn component_wise<F: Fn(Self::Scalar, Self::Scalar) -> Self::Scalar>(&self, rhs: &Self, f: F) -> Self { ... }
    fn map<F: Fn(Self::Scalar) -> O::Scalar, O: VectorN>(&self, f: F) -> O { ... }
    fn min_vec(&self, rhs: &Self) -> Self { ... }
    fn max_vec(&self, rhs: &Self) -> Self { ... }
    fn fold<T, F: Fn(T, Self::Scalar) -> T>(&self, acc: T, f: F) -> T { ... }
    fn all_comp_wise<F: Fn(Self::Scalar, Self::Scalar) -> bool>(&self, rhs: &Self, f: F) -> bool { ... }
    fn dot(&self, rhs: &Self) -> Self::Scalar { ... }
    fn length2(&self) -> Self::Scalar { ... }
}

Abstraction over vectors in different dimensions.

Associated Types

type Scalar: SpadeNum

The vector's internal scalar type.

Required Methods

fn from_value(value: Self::Scalar) -> Self

Creates a new vector with all compoenents set to a certain value.

fn dimensions() -> usize

The (fixed) number of dimensions of this vector trait.

Provided Methods

fn new() -> Self

Creates a new vector with all components initialized to zero.

fn component_wise<F: Fn(Self::Scalar, Self::Scalar) -> Self::Scalar>(&self, rhs: &Self, f: F) -> Self

Applies a binary operation component wise.

fn map<F: Fn(Self::Scalar) -> O::Scalar, O: VectorN>(&self, f: F) -> O

Maps an unary operation to all compoenents.

fn min_vec(&self, rhs: &Self) -> Self

Returns a new vector containing the minimum values of this and another vector (componentwise)

fn max_vec(&self, rhs: &Self) -> Self

Returns a new vector containing the maximum values of this and another vector (componentwise)

fn fold<T, F: Fn(T, Self::Scalar) -> T>(&self, acc: T, f: F) -> T

Fold operation over all vector components.

fn all_comp_wise<F: Fn(Self::Scalar, Self::Scalar) -> bool>(&self, rhs: &Self, f: F) -> bool

Checks if a property holds for all components.

fn dot(&self, rhs: &Self) -> Self::Scalar

Returns the vector's dot product.

fn length2(&self) -> Self::Scalar

Returns the vector's squared length.