Type Alias spade::handles::DirectedVoronoiEdge

pub type DirectedVoronoiEdge<'a, V, DE, UE, F> = DynamicHandleImpl<'a, V, DE, UE, F, DirectedVoronoiEdgeTag, InnerTag>;

A handle to a directed edge of the Voronoi diagram.

Several methods are defined to explore adjacent edges, faces and vertices:

Aliased Type

struct DirectedVoronoiEdge<'a, V, DE, UE, F> { /* private fields */ }

Implementations

impl<'a, V, DE, UE, F> DirectedVoronoiEdge<'a, V, DE, UE, F>

pub fn to(&self) -> VoronoiVertex<'a, V, DE, UE, F>

Returns the voronoi edge’s destination.

pub fn from(&self) -> VoronoiVertex<'a, V, DE, UE, F>

Returns the voronoi vertex from which this edge originates.

pub fn face(&self) -> VoronoiFace<'a, V, DE, UE, F>

Returns the Voronoi face to the left of this Voronoi edge

pub fn as_undirected(&self) -> UndirectedVoronoiEdge<'a, V, DE, UE, F>

Converts this directed edge handle into an undirected edge handle.

See also the handles module for more information.

pub fn as_delaunay_edge(&self) -> DirectedEdgeHandle<'a, V, DE, UE, F>

Returns the directed dual edge of the underlying Delaunay triangulation.

The dual edge is always orthogonal to to this edge.

Shows various inner voronoi edges (blue) and their dual Delaunay edges (orange)

pub fn rev(&self) -> Self

Returns this edge with its direction reversed.

pub fn next(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Returns the edge that is connected to this edge in counter clockwise order.

See also prev

pub fn prev(&self) -> DirectedVoronoiEdge<'a, V, DE, UE, F>

Returns the edge that is connected to this edge in clockwise order.

See also next

impl<'a, V, DE, UE, F> DirectedVoronoiEdge<'a, V, DE, UE, F>
where V: HasPosition,

pub fn direction_vector(&self) -> Point2<V::Scalar>

Returns a vector that is parallel to the voronoi edge.

This vector is obtained by rotating the dual Delaunay edge by 90° degree The returned vector is not necessarily normalized.