spade::handles
Type Alias FixedUndirectedEdgeHandle

pub type FixedUndirectedEdgeHandle = FixedHandleImpl<UndirectedEdgeTag, InnerTag>;

Fixed handle to an undirected edge.

See also the handles module.

Aliased Type

struct FixedUndirectedEdgeHandle { /* private fields */ }

Implementations

impl FixedUndirectedEdgeHandle

pub fn as_directed(&self) -> FixedDirectedEdgeHandle

Converts this directed edge into an undirected edge handle.

Any of the two directed edges may be returned.

See also FixedDirectedEdgeHandle::as_undirected

pub fn directed_edges(&self) -> [FixedDirectedEdgeHandle; 2]

Returns the two directed edges of this undirected edge in any order.