Type Definition spade::handles::VertexHandle

pub type VertexHandle<'a, V, DE = (), UE = (), F = ()> = DynamicHandleImpl<'a, V, DE, UE, F, VertexTag, InnerTag>;

Handle to a vertex of a triangulation.

Use this handle to retrieve the vertex position or its outgoing edges.

See also the handles module.

Implementations

impl<'a, V, DE, UE, F> VertexHandle<'a, V, DE, UE, F>where V: HasPosition,

pub fn position(&self) -> Point2<V::Scalar>

Returns the position of this vertex.

impl<'a, V, DE, UE, F> VertexHandle<'a, V, DE, UE, F>

pub fn out_edges( &self ) -> CircularIterator<'a, V, DE, UE, F, CCWEdgesNextBackFn>

Returns all directed edges going out of this vertex.

The edges are returned in counter clockwise order, beginning at an arbitrary edge.

A possible iteration order of v.out_edges()

Note: The returned iterator implements DoubleEndedIterator, allowing traversal in clockwise order.

pub fn out_edge(&self) -> Option<DirectedEdgeHandle<'a, V, DE, UE, F>>

Returns an outgoing edge of this vertex.

If the vertex has multiple outgoing edges, any of them is returned.

pub fn data(&self) -> &V

Returns the data associated with this vertex.

pub fn as_voronoi_face(&self) -> VoronoiFace<'a, V, DE, UE, F>

Returns the voronoi face that corresponds to this vertex of the Delaunay triangulation.