Struct spade::primitives::EdgeSideInfo

pub struct EdgeSideInfo<S> { /* fields omitted */ }

Yields information on which side of a line a point lies.

Methods

impl<S> EdgeSideInfo<S> where
S: SpadeNum,

pub fn is_on_left_side(&self) -> bool

Returns true if the query point lies on the left side of the query edge.

pub fn is_on_right_side(&self) -> bool

Returns true if the query point lies on the right side of the query edge.

pub fn is_on_left_side_or_on_line(&self) -> bool

Returns true if the query point lies on the left side or on the query edge.

pub fn is_on_right_side_or_on_line(&self) -> bool

Returns true if the query point lies on the right side or on the query edge.

pub fn is_on_line(&self) -> bool

Returns true if the query point lies on an infinite line going through the edge's start and end point.

pub fn reversed(&self) -> EdgeSideInfo<S>

Inverts this edge side information. If this information indicates the position of a point P relative to an edge A -> B, the inverted information will indicate the relative position of P relative to the edge B -> A.