Struct spade::primitives::SimpleEdge

pub struct SimpleEdge<V: PointN> {
    pub from: V,
    pub to: V,
}

An edge defined by it's two end points.

Fields
from: V

The edge's origin.

to: V

The edge's destination.

Methods

impl<V> SimpleEdge<V> where
V: PointN,

pub fn new(from: V, to: V) -> SimpleEdge<V>

Creates a new edge from from to to.

pub fn is_projection_on_edge(&self, query_point: &V) -> bool

Projects a point onto the infinite line going through the edge's start and end point and returns true if the projected points lies between from and to.

pub fn length2(&self) -> V::Scalar

Returns the edge's squared length.

impl<V> SimpleEdge<V> where
V: TwoDimensional,

pub fn side_query<K: DelaunayKernel<V::Scalar>>(
&self,
q: &V
) -> EdgeSideInfo<V::Scalar>

Determines on which side of this edge a given point lies.

Example:

use nalgebra::Point2;
use spade::kernels::TrivialKernel;
use spade::primitives::SimpleEdge;

let e = SimpleEdge::new(Point2::new(0f32, 0.), Point2::new(1., 1.));
assert!(e.side_query::<TrivialKernel>(&Point2::new(1.0, 0.0)).is_on_right_side());
assert!(e.side_query::<TrivialKernel>(&Point2::new(0.0, 1.0)).is_on_left_side());
assert!(e.side_query::<TrivialKernel>(&Point2::new(0.5, 0.5)).is_on_line());

pub fn intersects_edge_non_collinear<K>(&self, other: &SimpleEdge<V>) -> bool where
K: DelaunayKernel<V::Scalar>,

Checks if this and another edge intersect.

The edges must not be collinear. Also, true is returned if the edges just touch each other.

Panics

Panics if both lines are collinear.

impl<V> SimpleEdge<V> where
V: PointN,
V::Scalar: SpadeFloat,

pub fn nearest_point(&self, query_point: &V) -> V

Yields the nearest point on this edge.

pub fn projection_distance2(&self, query_point: &V) -> V::Scalar

Returns the squared distance of a given point to its projection onto the infinite line going through this edge's start and end point.

pub fn project_point(&self, query_point: &V) -> V::Scalar

Projects a point on this line and returns its relative position.

This method will return a value between 0. and 1. (linearly interpolated) if the projected point lies between self.from and self.to, a value close to zero (due to rounding errors) if the projected point is equal to self.from and a value smaller than zero if the projected point lies "before" self.from. Analogously, a value close to 1. or greater than 1. is returned if the projected point is equal to or lies behind self.to.