Struct spade::primitives::SimpleTriangle

pub struct SimpleTriangle<V: PointN> { /* fields omitted */ }

A triangle, defined by it's three points.

Methods

impl<V> SimpleTriangle<V> where
V: PointN,

pub fn new(v0: V, v1: V, v2: V) -> SimpleTriangle<V>

Checks if the given points are ordered counter clock wise.

pub fn vertices(&self) -> [&V; 3]

Returns the triangle's vertices.

impl<V: TwoDimensional> SimpleTriangle<V> where
V: TwoDimensional,

pub fn double_area(&self) -> V::Scalar

Returns the triangle's doubled area.

impl<V> SimpleTriangle<V> where
V: PointN,
V::Scalar: SpadeFloat,

pub fn nearest_point_on_edge(&self, pos: &V) -> V

Returns the nearest point lying on any of the triangle's edges.

impl<V> SimpleTriangle<V> where
V: TwoDimensional,
V::Scalar: SpadeFloat,

pub fn circumcenter(&self) -> V

Returns the position of the triangle's circumcenter.

pub fn barycentric_interpolation(&self, coord: &V) -> Point3<V::Scalar>

Returns the barycentric coordinates of a point.