Struct spade::DelaunayTriangulation

pub struct DelaunayTriangulation<V: HasPosition2D, K> where V::Vector: TwoDimensional {
    // some fields omitted
}

A 2D Delaunay triangulation.

A delaunay triangulation is a special triangulation of a set of points that fulfills some suitable properties for geometric operations like interpolation.

This triangulation works with Vector2-vectors from the cgmath and nalgebra package. Objects that are inserted into the triangulation have to implement the HasPosition2D trait.

Implementing delaunay triangulations is all about precision: the various geometric queries can fail if imprecise calculations are used (like native f32 / f64 operations), resulting in crashes at run time. To prevent those crashes, Spade offers a few "calculation kernels" that can fit the individual needs of an application. Refer to the following table and the documentation of each kernel for more information:

Vector types: 	When to use: 	Properties:

TrivialKernel 	recommended: f64, i64 also supported: f32, i32 	For i64 coordinates in the range of ± 100000. For f64 coordinates if performance is your main concern. 	Fastest performance. Can crash due to rounding (float types) and overflow (int types) issues.

FloatKernel 	f64 	Recommended for f64 coordinates. 	Still pretty fast. Uses adaptive precise arithmetic to prevent (theoretically) all crashes.

AdaptiveIntKernel 	i64 	For i64 coordinates with large value range. 	Slower than FloatKernel. Consider casting to floats and using those instead.

Creation

A triangulation can be created with <kernel_name>::new_triangulation(), e.g.

use spade::{DelaunayKernel, FloatKernel};
let mut triangulation = FloatKernel::new_triangulation();
// Once you insert the first piece of data, type inference is able to fully determine the
// triangulation's type - in this case, it contains nalgebra::Vector2<f64>.
triangulation.insert(nalgebra::Vector2::new(332f64, 123f64));

Also, Default is implemented, creating a triangulation with the trivial kernel.

Example

extern crate nalgebra;
extern crate spade;

use nalgebra::{Vector2};
use spade::{DelaunayTriangulation, DelaunayTriangle};

  let mut delaunay = DelaunayTriangulation::default();
  delaunay.insert(Vector2::new(0.0, 1.0));
  delaunay.insert(Vector2::new(0.0, -1.0));
  delaunay.insert(Vector2::new(1.0, 0.0));
  delaunay.insert(Vector2::new(-1.0, 0.0));
  delaunay.insert(Vector2::new(0.0, 0.0));
  for DelaunayTriangle(vertices) in delaunay.triangles() {
    println!("found triangle: {:?} -> {:?} -> {:?}", *vertices[0], *vertices[1], *vertices[2]);
  }
  for edge in delaunay.edges() {
    println!("found an edge: {:?} -> {:?}", *edge.from_handle(), *edge.to_handle());
  }

Iterating

A triangulation has three elements - vertices, edges and triangles - that can be iterated over. Use vertices() edges() and triangles() to call appropriate, non-mutating iterators.

Mutating

Adding vertices is always a legal operation. Removing vertices is not possible. Mutation of a vertex is possible with lookup_mut(..).

While VertexHandles are intended to be used for short lived iteration purposes, FixedVertexHandles can be made persistent and have a 'static lifetime. A VertexHandle can be transformed to its fixed variant by calling VertexHandle::fix(). Use handle or handle_mut to resolve a FixedVertexHandle.

A vertex position must not be changed after the vertex has been inserted.

Methods

impl<V, K> DelaunayTriangulation<V, K> where V: HasPosition2D, K: DelaunayKernel<V::Vector::Scalar>, V::Vector: TwoDimensional

fn new() -> DelaunayTriangulation<V, K>

Creates a new Delaunay triangulation.

Using this method directly can be a bit cumbersome due to type annotations, consider using Default::default() or <kernel name>::new_triangulation() for a DelaunayKernel if possible. Otherwise, you may need to specify the kernel like this:

use spade::{DelaunayTriangulation, FloatKernel};
let mut triangulation = DelaunayTriangulation::<_, FloatKernel>::new();
 

Usually, the omitted type (the triangulation's vertex type) can be inferred after insert is called.

fn handle(&self, handle: FixedVertexHandle) -> VertexHandle<V, K>

Creates a dynamic vertex handle from a fixed handle. May panic if the handle was not obtained from this triangulation.

fn handle_mut(&mut self, handle: FixedVertexHandle) -> &mut V

Returns a mutable reference to the vertex data referenced by a FixedVertexHandle. May panic if the handle was not obtained from this triangulation.

fn lookup(&self, point: &V::Vector) -> Option<VertexHandle<V, K>>

Checks if the triangulation contains an object with a given coordinate.

fn lookup_mut(&mut self, point: &V::Vector) -> Option<&mut V>

Checks if the triangulation contains an object and returns a mutable reference to it. Note that this will return a reference, while lookup(..) returns a vertex handle.

fn lookup_in_rect(&self, rect: &BoundingRect<V::Vector>) -> Vec<VertexHandle<V, K>>

Returns all vertices contained in a rectangle.

fn lookup_in_circle(&self, center: &V::Vector, radius2: &V::Vector::Scalar) -> Vec<VertexHandle<V, K>>

Returns all vertices contained in a circle.

fn num_vertices(&self) -> usize

Returns the number of vertices in this triangulation.

fn triangles(&self) -> DelaunayTriangleIterator<V, K>

Returns an iterator over all triangles.

fn edges(&self) -> AllEdgesIterator<V, K>

Returns an iterator over all edges.

fn vertices(&self) -> AllVerticesIterator<V, K>

Returns an iterator over all vertices.

fn get_position_in_triangulation(&self, point: &V::Vector) -> PositionInTriangulation<VertexHandle<V, K>>

Returns information about the location of a point in a triangulation.

fn insert(&mut self, t: V) -> FixedVertexHandle

Inserts a new vertex into the triangulation This operations runs in O(log n) on average, n denotes the number of vertices contained in the triangulation. If the point has already been contained in the triangulation, the old vertex is overwritten.

Returns a handle to the new vertex. Use this handle with DelaunayTriangulation::handle(..) to refer to the vertex.

impl<V, K> DelaunayTriangulation<V, K> where V: HasPosition2D, V::Vector::Scalar: SpadeFloat, K: DelaunayKernel<V::Vector::Scalar>, V::Vector: TwoDimensional

fn nn_interpolation<R, F: Fn(&V) -> R>(&self, point: &V::Vector, f: F) -> Option<R> where R: Mul<V::Vector::Scalar, Output=R> + Add<Output=R> + Sub<R>

Performs a natural neighbor interpolation for a given position.

Returns None if the triangulation has no triangles yet. Points outside of the convex hull will be interpolated as well. This operation runs in O(log n) for n inserted vertices.

Example

struct PointWithHeight {
  point: Vector2<f32>,
  height: f32,
}

impl HasPosition for PointWithHeight {
  type Vector = Vector2<f32>;
    fn position(&self) -> Vector2<f32> {
      self.point
    }
}

fn main() {
  let mut delaunay = DelaunayTriangulation::default();
  delaunay.insert(PointWithHeight { point: Vector2::new(0.0, 0.0), height: 5. });
  delaunay.insert(PointWithHeight { point: Vector2::new(1.0, 0.0), height: 0. });
  delaunay.insert(PointWithHeight { point: Vector2::new(0.0, 1.0), height: 0. });
  let lookup = Vector2::new(0.2, 0.2);
  // Interpolate the points height
  let interpolated = delaunay.nn_interpolation(&lookup, |p| p.height).unwrap();
  // and insert it afterwards.
  delaunay.insert(PointWithHeight { point: lookup, height: interpolated });
  // Data points themselves will always yield their own height
  assert_eq!(delaunay.nn_interpolation(&Vector2::new(0.0, 0.0), |p| p.height),
             Some(5.0));
}

fn estimate_normals<F, G, RV>(&mut self, f: &F, g: G) where F: Fn(&V) -> V::Vector::Scalar, G: Fn(&mut V, RV), RV: ThreeDimensional<Scalar=V::Vector::Scalar>

Estimates a normal for each vertex in the triangulation.

f must yield a "height" value for each vertex, g is a callback function that can be used to store the calculated normals.

 

struct PointWithHeight {
  point: Vector2<f32>,
  normal: Vector3<f32>,
  height: f32,
}

impl HasPosition for PointWithHeight {
  type Vector = Vector2<f32>;
    fn position(&self) -> Vector2<f32> {
      self.point
    }
}
impl PointWithHeight {
  fn new(point: Vector2<f32>, height: f32) -> PointWithHeight {
    PointWithHeight { point: point, height: height, normal: Vector3::new(0., 0., 0.) }
  }
}

fn main() {
  let mut delaunay = DelaunayTriangulation::default();
  // Insert some points here... (skipped)
  // Then, estimate all normals at once:
  delaunay.estimate_normals(&(|v: &PointWithHeight| v.height), &(|v: &mut PointWithHeight, n| v.normal = n));
   
  // And print them
  for vertex in delaunay.vertices() {
     println!("vertex: {:?}, normal: {:?}", vertex.position(), vertex.normal);
  }
}

fn estimate_normal<F, RV>(&self, v: FixedVertexHandle, f: &F) -> RV where F: Fn(&V) -> V::Vector::Scalar, RV: ThreeDimensional<Scalar=V::Vector::Scalar>

Estimates a normal value for a given vertex.

This assumes that the triangulation models some kind of height field, given by the function f. The normal is the weighted and normalized average of the normals of all triangles adjacent to the given vertex.

fn estimate_gradients<F, G>(&mut self, f: &F, g: &G) where F: Fn(&V) -> V::Vector::Scalar, G: Fn(&mut V, V::Vector)

Estimates gradients for all vertices in this triangulation.

f yields the value for which a gradient should be calculated, g can be used to store the gradient. See estimate_normals for a similar example.

Internally, the normal for each vertex is calculated first. Then, an appropriate gradient is calculated.

fn estimate_gradient<F, G>(&self, v: FixedVertexHandle, f: &F) -> V::Vector where F: Fn(&V) -> V::Vector::Scalar

Estimates and returns the gradient for a single vertex in this triangulation.

fn nn_interpolation_c1_sibson<F: Fn(&V) -> V::Vector::Scalar, G: Fn(&V) -> V::Vector>(&self, point: &V::Vector, f: F, g: G) -> Option<V::Vector::Scalar>

Interpolates a data point on this triangulation according to Sibson's c1 interpolant.

The interpolation given by nn_interpolation is not differentiable at the triangulation's data points. Sibson introduced another interpolation scheme that takes the gradient of each data point into account and offers an interpolation that is differentiable (c1) at the data points. The interpolation needs to know the gradients of the points natural neighbors, though. Spade can estimate them automatically, see estimate_gradient and estimate_gradients. The value that should be interpolated is given by f, the gradient of a vertex must be given by g.

Example

 
struct PointWithHeight {
  point: Vector2<f32>,
  gradient: Vector2<f32>,
  height: f32,
}

impl HasPosition for PointWithHeight {
  type Vector = Vector2<f32>;
    fn position(&self) -> Vector2<f32> {
      self.point
    }
}

impl PointWithHeight {
  fn new(point: Vector2<f32>, height: f32) -> PointWithHeight {
    // Initialize the gradient to any value since it will be overwritten
    PointWithHeight { point: point, height: height, gradient: Vector2::new(0., 0.) }
  }
}

fn main() {
  let mut delaunay = DelaunayTriangulation::default();
  // Insert some points here...
  // Estimate all gradients and store them:
  delaunay.estimate_gradients(&(|v: &PointWithHeight| v.height),
                              &(|v: &mut PointWithHeight, g| v.gradient = g));
   
  // Now we can use the gradients for interpolation:
 let interpolated = delaunay.nn_interpolation_c1_sibson(
     &Vector2::new(0.5, 0.2), |v| v.height, |v| v.gradient);
 println!("interpolated: {}", interpolated.unwrap());
}